getting-started-web-api/requirements.txt
Django>=4.2
psycopg2-binary>=2.9 
djangorestframework
django-oauth-toolkit>=2.0.0
django-cors-headers>=4.0.0
requests>=2.28.0
python-jose[cryptography]>=3.3.0 